Fix code review issues: DDL dedup, port type, test completeness
- Move DDLBintrailServers/DDLBintrailServerChanges to internal/serverid
  as exported constants — single source of truth; init.go and testutil
  both reference serverid.DDL* instead of each maintaining their own copy
- Change port column from INT UNSIGNED to SMALLINT UNSIGNED (matches TCP
  port range and uint16 in Go) in serverid.go, migrations/001_create_tables.sql
- Add args assertion to TestBuildQueryViaGlob: verify LIMIT ? clause and
  args=[50] so the test is a complete positive-path check
