+++ cmd/bintrail/init.go
@@ -15,6 +15,8 @@ import (
 	"github.com/aws/aws-sdk-go-v2/service/s3/types"
 	"github.com/go-sql-driver/mysql"
 	"github.com/spf13/cobra"
+
+	"github.com/bintrail/bintrail/internal/serverid"
 )
 
 var initCmd = &cobra.Command{
@@ -490,28 +492,10 @@ const ddlIndexState = `CREATE TABLE IF NOT EXISTS index_state (
     error_message  TEXT     DEFAULT NULL
 ) ENGINE=InnoDB`
 
-const ddlBintrailServers = `CREATE TABLE IF NOT EXISTS bintrail_servers (
-    bintrail_id       CHAR(36)     NOT NULL,
-    server_uuid       CHAR(36)     NOT NULL,
-    host              VARCHAR(255) NOT NULL,
-    port              INT UNSIGNED NOT NULL DEFAULT 3306,
-    username          VARCHAR(255) NOT NULL,
-    created_at        TIMESTAMP    NOT NULL DEFAULT CURRENT_TIMESTAMP,
-    updated_at        TIMESTAMP    NOT NULL DEFAULT CURRENT_TIMESTAMP ON UPDATE CURRENT_TIMESTAMP,
-    decommissioned_at TIMESTAMP    NULL DEFAULT NULL,
-    PRIMARY KEY (bintrail_id),
-    INDEX idx_server_uuid  (server_uuid),
-    INDEX idx_host_port_user (host, port, username)
-) ENGINE=InnoDB`
-
-const ddlBintrailServerChanges = `CREATE TABLE IF NOT EXISTS bintrail_server_changes (
-    id            BIGINT UNSIGNED NOT NULL AUTO_INCREMENT,
-    bintrail_id   CHAR(36)        NOT NULL,
-    field_changed ENUM('server_uuid','host','port','username') NOT NULL,
-    old_value     VARCHAR(255)    NOT NULL,
-    new_value     VARCHAR(255)    NOT NULL,
-    detected_at   TIMESTAMP       NOT NULL DEFAULT CURRENT_TIMESTAMP,
-    PRIMARY KEY (id),
-    INDEX idx_bintrail_id (bintrail_id),
-    INDEX idx_detected_at (detected_at)
-) ENGINE=InnoDB`
+// ddlBintrailServers and ddlBintrailServerChanges are the canonical DDL
+// statements for the server identity tables. They live in internal/serverid
+// so that testutil and init share a single source of truth.
+var (
+	ddlBintrailServers      = serverid.DDLBintrailServers
+	ddlBintrailServerChanges = serverid.DDLBintrailServerChanges
+)
