+++ internal/testutil/testutil.go
@@ -12,6 +12,8 @@ import (
 	"testing"
 
 	_ "github.com/go-sql-driver/mysql"
+
+	"github.com/bintrail/bintrail/internal/serverid"
 )
 
 // dbCounter ensures unique database names across calls within the same test.
@@ -194,31 +196,8 @@ func InitIndexTables(t *testing.T, db *sql.DB) {
 		CONSTRAINT single_row CHECK (id = 1)
 	) ENGINE=InnoDB`)
 
-	MustExec(t, db, `CREATE TABLE IF NOT EXISTS bintrail_servers (
-		bintrail_id       CHAR(36)     NOT NULL,
-		server_uuid       CHAR(36)     NOT NULL,
-		host              VARCHAR(255) NOT NULL,
-		port              INT UNSIGNED NOT NULL DEFAULT 3306,
-		username          VARCHAR(255) NOT NULL,
-		created_at        TIMESTAMP    NOT NULL DEFAULT CURRENT_TIMESTAMP,
-		updated_at        TIMESTAMP    NOT NULL DEFAULT CURRENT_TIMESTAMP ON UPDATE CURRENT_TIMESTAMP,
-		decommissioned_at TIMESTAMP    NULL DEFAULT NULL,
-		PRIMARY KEY (bintrail_id),
-		INDEX idx_server_uuid    (server_uuid),
-		INDEX idx_host_port_user (host, port, username)
-	) ENGINE=InnoDB`)
-
-	MustExec(t, db, `CREATE TABLE IF NOT EXISTS bintrail_server_changes (
-		id            BIGINT UNSIGNED NOT NULL AUTO_INCREMENT,
-		bintrail_id   CHAR(36)        NOT NULL,
-		field_changed ENUM('server_uuid','host','port','username') NOT NULL,
-		old_value     VARCHAR(255)    NOT NULL,
-		new_value     VARCHAR(255)    NOT NULL,
-		detected_at   TIMESTAMP       NOT NULL DEFAULT CURRENT_TIMESTAMP,
-		PRIMARY KEY (id),
-		INDEX idx_bintrail_id (bintrail_id),
-		INDEX idx_detected_at (detected_at)
-	) ENGINE=InnoDB`)
+	MustExec(t, db, serverid.DDLBintrailServers)
+	MustExec(t, db, serverid.DDLBintrailServerChanges)
 }
 
 // InsertEvent inserts a single event into binlog_events using raw SQL.
